Move DownloadedObject into new file
