+++ packages/evo-objects/src/evo/objects/client/api_client.py
@@ -11,17 +11,16 @@
 
 from __future__ import annotations
 
-from collections.abc import AsyncIterator, Iterator, Sequence
+from collections.abc import AsyncIterator, Sequence
 from uuid import UUID
 
 from evo import logging
 from evo.common import APIConnector, BaseAPIClient, HealthCheckType, ICache, Page, ServiceHealth
 from evo.common.data import EmptyResponse, Environment, OrderByOperatorEnum
-from evo.common.io.exceptions import DataNotFoundError
 from evo.common.utils import get_service_health, parse_order_by
 
 from .. import parse
-from ..data import ObjectMetadata, ObjectOrderByEnum, ObjectSchema, ObjectVersion, OrgObjectMetadata, Stage
+from ..data import ObjectMetadata, ObjectOrderByEnum, ObjectVersion, OrgObjectMetadata, Stage
 from ..endpoints import MetadataApi, ObjectsApi, StagesApi
 from ..endpoints.models import (
     GeoscienceObject,
@@ -32,59 +31,11 @@ from ..endpoints.models import (
 from ..exceptions import ObjectUUIDError
 from ..io import ObjectDataDownload, ObjectDataUpload
 from ..utils import ObjectDataClient
+from .object_client import DownloadedObject
 
 logger = logging.getLogger("object.client")
 
-__all__ = [
-    "DownloadedObject",
-    "ObjectAPIClient",
-]
-
-
-class DownloadedObject:
-    """A downloaded geoscience object."""
-
-    def __init__(
-        self, object_: GeoscienceObject, metadata: ObjectMetadata, urls_by_name: dict[str, str], connector: APIConnector
-    ) -> None:
-        self._object = object_
-        self._metadata = metadata
-        self._urls_by_name = urls_by_name
-        self._connector = connector
-
-    @property
-    def schema(self) -> ObjectSchema:
-        """The schema of the object."""
-        return self._metadata.schema_id
-
-    @property
-    def metadata(self) -> ObjectMetadata:
-        """The metadata of the object."""
-        return self._metadata
-
-    def as_dict(self) -> dict:
-        """Get this object as a dictionary."""
-        return self._object.model_dump(mode="python", by_alias=True)
-
-    def prepare_data_download(self, data_identifiers: Sequence[str | UUID]) -> Iterator[ObjectDataDownload]:
-        """Prepare to download multiple data files from the geoscience object service, for this object.
-
-        Any data IDs that are not associated with the requested object will raise a DataNotFoundError.
-
-        :param data_identifiers: A list of sha256 digests or UUIDs for the data to be downloaded.
-
-        :return: An iterator of data download contexts that can be used to download the data.
-
-        :raises DataNotFoundError: If any requested data ID is not associated with this object.
-        """
-        try:
-            filtered_urls_by_name = {str(name): self._urls_by_name[str(name)] for name in data_identifiers}
-        except KeyError as exc:
-            raise DataNotFoundError(f"Unable to find the requested data: {exc.args[0]}") from exc
-        for ctx in ObjectDataDownload._create_multiple(
-            connector=self._connector, metadata=self._metadata, urls_by_name=filtered_urls_by_name
-        ):
-            yield ctx
+__all__ = ["ObjectAPIClient"]
 
 
 class ObjectAPIClient(BaseAPIClient):
