+++ packages/evo-objects/src/evo/objects/client/object_client.py
@@ -1,0 +1,73 @@
+#  Copyright © 2025 Bentley Systems, Incorporated
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#      http://www.apache.org/licenses/LICENSE-2.0
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+from __future__ import annotations
+
+from collections.abc import Iterator, Sequence
+from uuid import UUID
+
+from evo import logging
+from evo.common import APIConnector
+from evo.common.io.exceptions import DataNotFoundError
+
+from ..data import ObjectMetadata, ObjectSchema
+from ..endpoints.models import GeoscienceObject
+from ..io import ObjectDataDownload
+
+__all__ = ["DownloadedObject"]
+
+logger = logging.getLogger("object.client")
+
+
+class DownloadedObject:
+    """A downloaded geoscience object."""
+
+    def __init__(
+        self, object_: GeoscienceObject, metadata: ObjectMetadata, urls_by_name: dict[str, str], connector: APIConnector
+    ) -> None:
+        self._object = object_
+        self._metadata = metadata
+        self._urls_by_name = urls_by_name
+        self._connector = connector
+
+    @property
+    def schema(self) -> ObjectSchema:
+        """The schema of the object."""
+        return self._metadata.schema_id
+
+    @property
+    def metadata(self) -> ObjectMetadata:
+        """The metadata of the object."""
+        return self._metadata
+
+    def as_dict(self) -> dict:
+        """Get this object as a dictionary."""
+        return self._object.model_dump(mode="python", by_alias=True)
+
+    def prepare_data_download(self, data_identifiers: Sequence[str | UUID]) -> Iterator[ObjectDataDownload]:
+        """Prepare to download multiple data files from the geoscience object service, for this object.
+
+        Any data IDs that are not associated with the requested object will raise a DataNotFoundError.
+
+        :param data_identifiers: A list of sha256 digests or UUIDs for the data to be downloaded.
+
+        :return: An iterator of data download contexts that can be used to download the data.
+
+        :raises DataNotFoundError: If any requested data ID is not associated with this object.
+        """
+        try:
+            filtered_urls_by_name = {str(name): self._urls_by_name[str(name)] for name in data_identifiers}
+        except KeyError as exc:
+            raise DataNotFoundError(f"Unable to find the requested data: {exc.args[0]}") from exc
+        for ctx in ObjectDataDownload._create_multiple(
+            connector=self._connector, metadata=self._metadata, urls_by_name=filtered_urls_by_name
+        ):
+            yield ctx
